+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 - Added expiry TTL to `tykcommon`, data expiry headers will be added to all analytics records, set `expire_analytics_after` to `0` to have data live indefinetely (currently 100 years), set to anything above zero for data in MongoDB to be removed after x seconds. **requirement**: You must create an expiry TTL index on the tyk_analytics collection manually (http://docs.mongodb.org/manual/tutorial/expire-data/). If you do not wish mongo to manage data warehousing at all, simply do not create the index.
 - Added a JS Virtual Machine so dynamic JS middleware can be run PRE and POST middleware chain
+- Added a global JS VM
+- Added an `eh_dynamic_handler` event handler type that runs JS event handlers
  
 
 # v1.3: